feat: create modules for E2E Vuln operations
