+++ deps/wazuh_testing/wazuh_testing/end_to_end/regexes/regex.yaml
