+++ deps/wazuh_testing/wazuh_testing/end_to_end/services.py
@@ -1,0 +1,4 @@
+def control_environment(host_manager, operation, group_list):
+    for group in group_list:
+        for host in host_manager.get_group_hosts(group):
+            host_manager.handle_wazuh_services(host, operation)
