+++ deps/wazuh_testing/wazuh_testing/modules/vulnerability_detector/event_monitor.py
@@ -167,7 +167,7 @@ def check_vulnerability_full_scan_end(log_monitor=None, agent_id='000'):
         agent_id (str): Agent ID.
     """
     check_vuln_detector_event(file_monitor=log_monitor, timeout=vd.T_40,
-                              callback=f"Finished vulnerability assessment for agent '{agent_id}'",
+                              callback=rf"Finished vulnerability assessment for agent \\'{agent_id}\\'",
                               error_message='No full scan end has been detected in the log.')
