+++ provisioning/roles/createInventory.yaml
@@ -3,5 +3,5 @@
   tasks:
     - name: Template a file
       ansible.builtin.template:
-        src: "{{ source_template }}"
+        src:  "{{ source_template }}"
         dest: "{{ dest_inventory}}"
