+++ tests/end_to_end/test_vulnerability_detector/cases.yaml
@@ -8,53 +8,55 @@
           package:
             centos:
               amd64: https://downloads.rclone.org/v1.49.5/rclone-v1.49.5-linux-amd64.rpm
-              arm: null
+              arm: https://download.postgresql.org/pub/repos/yum/11/redhat/rhel-7-aarch64/postgresql11-libs-11.17-2PGDG.rhel7.aarch64.rpm
             ubuntu:
               amd64: https://downloads.rclone.org/v1.49.5/rclone-v1.49.5-linux-amd64.deb
-              arm: null
+              arm: https://grafana.com/grafana/download/8.5.5?edition=enterprise&platform=arm
             windows:
               amd64: https://get.videolan.org/vlc/3.0.6/win64/vlc-3.0.6-win64.exe
             macos:
               amd64: https://nodejs.org/dist/v17.0.1/node-v17.0.1.pkg
-
-# {"timestamp":"2023-11-14T10:52:41.932+0000","rule":{"level":10,"description":"CVE-2020-28924 affects rclone","id":"23505","firedtimes":392,"mail":false,"groups":["vulnerability-detector"],"gdpr":["IV_35.7.d"],"pci_dss":["11.2.1","11.2.3"],"tsc":["CC7.1","CC7.2"]},"agent":{"id":"002","name":"ip-172-31-12-122.ec2.internal","ip":"172.31.12.122"},"manager":{"name":"ip-172-31-6-24"},"id":"1699959161.19604197","cluster":{"name":"wazuh","node":"master"},"decoder":{"name":"json"},"data":{"vulnerability":{"package":{"name":"rclone","version":"1.49.5-1","architecture":"x86_64","condition":"Package less than 1.53.3"},"cvss":{"cvss2":{"vector":{"attack_vector":"network","access_complexity":"low","authentication":"none","confidentiality_impact":"partial","integrity_impact":"none","availability":"none"},"base_score":"5","exploitability_score":"10","impact_score":"2.900000"},"cvss3":{"vector":{"attack_vector":"network","access_complexity":"low","privileges_required":"none","user_interaction":"none","scope":"unchanged","confidentiality_impact":"high","integrity_impact":"none","availability":"none"},"base_score":"7.500000","exploitability_score":"3.900000","impact_score":"3.600000"}},"cve":"CVE-2020-28924","title":"CVE-2020-28924 affects rclone","rationale":"An issue was discovered in Rclone before 1.53.3. Due to the use of a weak random number generator, the password generator has been producing weak passwords with much less entropy than advertised. The suggested passwords depend deterministically on the time the second rclone was started. This limits the entropy of the passwords enormously. These passwords are often used in the crypt backend for encryption of data. It would be possible to make a dictionary of all possible passwords with about 38 million entries per password length. This would make decryption of secret material possible with a plausible amount of effort. NOTE: all passwords generated by affected versions should be changed.","severity":"High","published":"2020-11-19","updated":"2022-04-26","cwe_reference":"CWE-331","status":"Active","type":"PACKAGE","references":["https://github.com/rclone/rclone/issues/4783","https://lists.fedoraproject.org/archives/list/package-announce@lists.fedoraproject.org/message/UJIFT24Q6EFXLQZ24AER2QGFFZLMIPCD/","https://rclone.org/downloads/","https://security.gentoo.org/glsa/202107-14","https://nvd.nist.gov/vuln/detail/CVE-2020-28924"],"assigner":"cve@mitre.org"}},"location":"vulnerability-detector"}
-
-    check_agent_alert_indexer:
-        centos:
-          amd64:
-            - event: syscollector_install_package_alert_yum
-              parameters:
-                PACKAGE_NAME: "rclone"
-                PACKAGE_VERSION: "1.49.5"
-          arm:
-            - event: syscollector_install_package_alert_yum
-              parameters:
-                PACKAGE_NAME: "rclone"
-                PACKAGE_VERSION: "1.49.5"
-        ubuntu:
-          amd64:
-            - event: syscollector_install_package_alert_apt
-              parameters:
-                PACKAGE_NAME: "rclone"
-                PACKAGE_VERSION: "1.49.5"
-          arm:
-            - event: syscollector_install_package_alert_apt
-              parameters:
-                PACKAGE_NAME: "rclone"
-                PACKAGE_VERSION: "1.49.5"
-        windows:
-          amd64:
-            - event: syscollector_install_package_alert_apt
-              parameters:
-                PACKAGE_NAME: "VideoLAN"
-                PACKAGE_VERSION: "3.0.6"
-        macos:
-          amd64:
-            - event: syscollector_install_package_alert_apt
-              parameters:
-                PACKAGE_NAME:
-                PACKAGE_VERSION:
-
+      - check_agent_vulnerability:
+            parameters:
+              alert_indexed: False
+              api: False
+              alert: True
+              state_indice: False
+            vulnerability_data:
+              centos:
+                amd64:
+                    PACKAGE_NAME: "rclone"
+                    PACKAGE_VERSION: "1.49.5"
+                    CVE: CVE-2020-28924
+                    ARCHITECTURE: 'x86_64'
+                arm:
+                    PACKAGE_NAME: "postgresql11-libs"
+                    PACKAGE_VERSION: "11.17.2"
+                    CVE: CVE-2020-28924
+                    ARCHITECTURE: 'arm64'
+              ubuntu:
+                amd64:
+                    PACKAGE_NAME: "rclone"
+                    PACKAGE_VERSION: "1.49.5"
+                    CVE: CVE-2020-28924
+                    ARCHITECTURE: 'x86_64'
+                arm:
+                    PACKAGE_NAME: "grafana"
+                    PACKAGE_VERSION: "8.5.5"
+                    CVE: CVE-2020-28924
+                    ARCHITECTURE: 'arm64'
+              windows:
+                amd64:
+                    PACKAGE_NAME: "vlc"
+                    PACKAGE_VERSION: "3.0.6"
+                    CVE: CVE-2020-28924
+                    ARCHITECTURE: 'x86_64'
+              macos:
+                amd64:
+                    PACKAGE_NAME: "node"
+                    PACKAGE_VERSION: "17.0.1"
+                    CVE: CVE-2020-28924
+                    ARCHITECTURE: 'x86_64'
 
     # check_alerts:
     #     centos:
@@ -91,19 +93,19 @@
     #           parameters:
     #             PACKAGE_NAME:
     #             PACKAGE_VERSION:
-  teardown:
-    - remove_package:
-        package:
-          centos:
-            amd: rclone
-            arm: rclone
-          ubuntu:
-            amd: rclone
-            arm: rclone
-          windows:
-            amd64: C:\\\\Program Files\\\\VideoLAN\\\\VLC\\\\uninstall.exe
-          macos:
-            amd64:
+  # teardown:
+    # - remove_package:
+    #     package:
+    #       centos:
+    #         amd: rclone
+    #         arm: rclone
+    #       ubuntu:
+    #         amd: rclone
+    #         arm: rclone
+    #       windows:
+    #         amd64: C:\\\\Program Files\\\\VideoLAN\\\\VLC\\\\uninstall.exe
+    #       macos:
+    #         amd64:
 
 # - case: "Updating a vulnerable package that remains vulnerable to the same CVE"
 #   id: "update_vuln_package_remain_vulnerable"
